+++ protocols/v2/channels-sv2/src/client/standard.rs
@@ -330,7 +330,7 @@ impl<'a> StandardChannel<'a> {
         // check if a block was found
         if network_target.is_met_by(hash) {
             self.share_accounting.update_share_accounting(
-                self.target.difficulty_float() as u64,
+                self.target.difficulty_float(),
                 share.sequence_number,
                 hash.to_raw_hash(),
             );
@@ -344,7 +344,7 @@ impl<'a> StandardChannel<'a> {
             }
 
             self.share_accounting.update_share_accounting(
-                self.target.difficulty_float() as u64,
+                self.target.difficulty_float(),
                 share.sequence_number,
                 hash.to_raw_hash(),
             );
