+++ roles/jd-client/src/lib/channel_manager/downstream_message_handler.rs
@@ -974,7 +974,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                                new_shares_sum: share_accounting.get_share_work_sum(),
+                                new_shares_sum: share_accounting.get_share_work_sum() as u64,
                             };
                             info!("SubmitSharesStandard on downstream channel: {} ✅", success);
                             messages.push((downstream.downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -1006,7 +1006,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                            new_shares_sum: share_accounting.get_share_work_sum(),
+                            new_shares_sum: share_accounting.get_share_work_sum() as u64,
                         };
                         messages.push((
                             downstream.downstream_id,
@@ -1169,7 +1169,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                                new_shares_sum: share_accounting.get_share_work_sum(),
+                                new_shares_sum: share_accounting.get_share_work_sum() as u64,
                             };
                             info!("SubmitSharesExtended on downstream channel: {} ✅", success);
                             messages.push((downstream.downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -1199,7 +1199,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                            new_shares_sum: share_accounting.get_share_work_sum(),
+                            new_shares_sum: share_accounting.get_share_work_sum() as u64,
                         };
                         is_downstream_share_valid = true;
                         messages.push((
