+++ roles/pool/src/lib/channel_manager/mining_message_handler.rs
@@ -536,7 +536,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                                new_shares_sum: share_accounting.get_share_work_sum(),
+                                new_shares_sum: share_accounting.get_share_work_sum() as u64,
                             };
                             info!("SubmitSharesStandard: {} ✅", success);
                             messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -568,7 +568,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                            new_shares_sum: share_accounting.get_share_work_sum(),
+                            new_shares_sum: share_accounting.get_share_work_sum() as u64,
                         };
                         messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
                     }
@@ -693,7 +693,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                                new_shares_sum: share_accounting.get_share_work_sum(),
+                                new_shares_sum: share_accounting.get_share_work_sum() as u64,
                             };
                             info!("SubmitSharesExtended: {} ✅", success);
                             messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -724,7 +724,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_shares_accepted(),
-                            new_shares_sum: share_accounting.get_share_work_sum(),
+                            new_shares_sum: share_accounting.get_share_work_sum() as u64,
                         };
                         messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
                     }
